Added information about the batch file
I have let users know what the batch file is and why it is there. The batch file just needs to be in the /videos folder as that is where is is currently coded to go. This can always be changed but at the moment it means that jumpcutter.py, jumpcutter.bat and ffmpeg.exe all have to be in the /videos file to run